modified dipole function to provide low memory, slow option if initial attempt fails... also moved analysis and geometry functions into two separate files for better organization
